refactor: split out modules
