+++ src/uvtrick/__init__.py
@@ -1,0 +1,2 @@
+from .core import *
+from .trick import uvtrick
